requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Cors
python-dotenv
Werkzeug
